Refactor: created package parsing.telegram and rename utils to models
